Update CODE.sql
The output column should be salary_diff instead of differences.

"The difference between the sum of the highest salaries and the sum of lowest salaries will give you the information you want. So, given the table employees, write a select statement as follows: The table should contain only one column salary_diff and only one row, which will contain the difference between sum of the highest and the sum of lowest salaries. If employees table is empty, the difference should be equal 0."